backend/requirements.txt
openai>=1.0.0
pandas>=1.5.0
numpy>=1.20.0
scikit-learn>=1.0.0
python-dotenv>=0.19.0
tqdm>=4.65.0
